add 32-aligned buffer type
This change adds `Buf32`, a heap-allocated 32-byte aligned buffer
type. It is meant to guarantee that

- buffers (and pointers to them) are aligned to a 32-byte boundary,
  and
- buffers have a byte length multiple of 32.

These are requirements for a lot of functions in libogc that take
buffers as parameters, so this eases the burden of meeting them when
writing code for them.

It currently does not grow or reallocate. This may or may not change
based on the needs of libogc and how it uses the buffer.

It's theoretically possible to implement Clone for it, but I just
haven't gotten around to it yet.